Use x_AddParameterToProgramOptions<Int8> to process GIs. JIRA CXX-8347
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93393 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/blast/api/search_strategy.cpp
@@ -445,9 +445,9 @@ void CExportStrategy::x_Process_SearchDb(CRef<CSearchDatabase> & db)
     if (!gi_list_limit.empty())
     {
 		#ifdef NCBI_STRICT_GI
-       		vector<int> int_gi_list;
+       		vector<Int8> int_gi_list;
        		ITERATE ( vector<TGi>, it, gi_list_limit ) {
-       			int_gi_list.push_back(GI_TO( int, *it));
+       			int_gi_list.push_back(GI_TO(Int8, *it));
        		}
        		x_AddParameterToProgramOptions(CBlast4Field::Get(eBlastOpt_GiList), int_gi_list);
 		#else
@@ -473,9 +473,9 @@ void CExportStrategy::x_Process_SearchDb(CRef<CSearchDatabase> & db)
     if (!neg_gi_list.empty())
     {
 		#ifdef NCBI_STRICT_GI
-       		vector<int> int_gi_list;
+       		vector<Int8> int_gi_list;
        		ITERATE ( vector<TGi>, it, neg_gi_list) {
-       			int_gi_list.push_back(GI_TO( int, *it));
+       			int_gi_list.push_back(GI_TO(Int8, *it));
        		}
        		x_AddParameterToProgramOptions(CBlast4Field::Get(eBlastOpt_NegativeGiList), int_gi_list);
 		#else
